[cfile] make WriterOptions* a constant member field
Since I'm preparing a series of patches to update files under
src/kudu/cfile, this patch does a minor cleanup on the files,
reflecting on the fact the WriteOptions* pointer field isn't
changing in all of the affected classes, and the kMaxHeaderSize
constant isn't used.

This patch doesn't contain any functional modifications.

Change-Id: I9a209194a15d9d49bbd42949223651b0fd17c7c0
Reviewed-on: http://gerrit.cloudera.org:8080/21631
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>