[tool] standardize the output of rebalance report tool
I discovered that in environments with authentication
enabled, if the kinit operation is not performed and
the 'kudu cluster rebalance §ma -report_only' command
is used to query the rebalance progress, an output
similar to
'I20240812 19:23:29.041643 15510 tool_action_cluster.cc:275] no version information from some servers; not rebalancing single-replica tablets as the result an empty cluster'
is returned. This output can mislead our users.

In this patch, I have fixed this unauthorized error output
and added corresponding unit tests to verify the new logic.

Change-Id: If844be1f7ccd156582f1ffa7137c8a711c49178d
Reviewed-on: http://gerrit.cloudera.org:8080/21672
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>