[ranger] enhance the robustness of key retrieval
In real-world scenarios where encryption keys are generated using
Ranger, we might encounter the following error when starting the
cluster:
'master_main.cc:42] Remote error: RunMasterServer() failed: Could not create new FS layout: unable to create instance metadata: failed to generate server key: HTTP 403'.
This error can be resolved by simply restarting without making any
changes.
Upon investigation, it was found that the keys requested from Ranger
have an effective period of 30 seconds after adding a new user, as
referenced in [1].
To enhance the robustness of the Kudu code, I have added a retry
mechanism for key retrieval in this patch to mitigate the impact of
Ranger user activation periods on the startup process.

Since only retry logic was added, no new unit tests were introduced.
However, I still verified the success rate of the new patch in a
real-world installation, and it reached 100%, which is a significant
improvement compared to the previous 50%.

[1]https://github.com/apache/ranger/blob/4e365456f6533ee5515c5070c92e355198922c81/agents-common/src/main/java/org/apache/ranger/plugin/util/PolicyRefresher.java#L92

Change-Id: I1fd3263ad6ba6d8e444036bb7d2158986098cb4b
Reviewed-on: http://gerrit.cloudera.org:8080/21673
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: KeDeng <kdeng.ke@gmail.com>