[rpc] instantiate DiagnosticSocket only if sock_diag available
In the review feedback on [1], Yingchun Lai pointed to the fact that
the DiagnosticSocket field would be instantiated in AcceptorPool even if
the sock_diag netlink facility was not available.  An alternative
approach would be not instantiating it at all in that case.  This patch
addresses the point.

This is a follow-up to d27603ec7ea6adb4e89c25a37ab1fed314011a47.

[1] http://gerrit.cloudera.org:8080/21722

Change-Id: Ic70b6cca94b50f070efbc159a767d207c3a7b230
Reviewed-on: http://gerrit.cloudera.org:8080/21727
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>