[java][build] KUDU-3610: kudu-client.jar dependencies
As reported in KUDU-3610 Impala build fails with the newly generated
kudu-client.jar pom.xml.

The issue is in the pom generation itself [1], that is being done by
the 'maven-publish' plugin (which is necessary for gradle 7+ due to
the old publish plugin being deprecated).

This means that the content of the jars are correct, however the
generated pom.xml lists some dependencies in incorrect (maven) scopes.

The fix this change proposes is by editing the generated pom xml
structure, so that the shaded dependencies are removed from the
generated pom. This was previously done by the shadow plugin, however
due to API changes this is no longer possible and has to be done in
maven-publish plugin.

Additional manual testing (building Impala) was carried out that
confirmed that the pom is now correct.

[1] https://discuss.gradle.org/t/maven-publish-plugin-generated-pom-making-dependency-scope-runtime/7494

Change-Id: I607d7c62c585b2aa1bee4ebf7907309e130092da
Reviewed-on: http://gerrit.cloudera.org:8080/21731
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 9befb1188ec2c8f75bf81f9cd5b17fec2db8c19b)
Reviewed-on: http://gerrit.cloudera.org:8080/21995
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>