[tablet] report per-thread status in mt-tablet-test
With this patch, mt-tablet-test now records per-thread status
and reports it in the very end of the scenario instead of crashing
on the first sight of non-OK status in a per-thread routine.
This patch also contains other minor modifications to make the code
in mt-tablet-test more robust.  This is in the context of KUDU-3465.

The motivation behind this update is to make sure that the sanitizers
(UBSAN/ASAN/TSAN etc) have a chance to report on an issue, if any.

Change-Id: I2e2cb8c3bbc9b12c5d45e603de1b0aeb14191f12
Reviewed-on: http://gerrit.cloudera.org:8080/21776
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>