[RowSetMetadata] fix const-correctness of a few methods
In review feedback for [1] it was mentioned to address const-correctness
of a few methods in the RowSetMetadata class.  This patch does exactly
so, also bringing the code up to date with the current C++ style guide.
One method in FsManager was also updated because it's related to these
changes in RowSetMetadata.

This patch doesn't contain any functional modifications.

[1] https://gerrit.cloudera.org/#/c/21779/

Change-Id: Idf9fae19cf77795730b1e58b98617bb29ed3552e
Reviewed-on: http://gerrit.cloudera.org:8080/21826
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
Tested-by: Alexey Serbin <alexey@apache.org>