[tablet] reduce lock contention in OpTracker::Release()
Since MemTracker::Release() runs tcmalloc's GC-ing which might
block for some time under heavy concurrent activity, it makes sense
to call it outside of the critical section that guards the map
with pending operations for the tablet.  Also, avoid double lookups
in that map, re-using the iterator for erasing the element.

Change-Id: Ic5e4298ccabe2f70500765d63083cc817b7a5f95
Reviewed-on: http://gerrit.cloudera.org:8080/21856
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>