[tests] fix failing TestWebUIDoesNotCrashCluster
Before this patch, the TestWebUIDoesNotCrashCluster scenario of the
webserver-stress-itest was failing if running in an environment where
KUDU_HOME was set and pointed to the correct location of the Kudu
git workspace.  Also, the webserver-stress-itest didn't look like a good
place for such a test.

This is a follow-up to b2514b98c7639c2a328a4dddd8fd9213428b0d4d.

Change-Id: I59b6ef64b522540620bdc622156b3516ada23f5d
Reviewed-on: http://gerrit.cloudera.org:8080/21863
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>