[dynamic_multi_master-test] reduce the amount of generated data
When I ran all C++ tests via ctest on one node having about 25GiBytes of
available disk space, it unexpectedly ran out of space.  The culprit was
dynamic_multi_master-test: the scenarios from the suite keep their test
workloads running while waiting for events that might take a long time
to happen, especially on slow or very busy machines.

This patch updates the configuration of the TestWorkload objects to use
a single writer thread (the default is four) in the TestBasicAddition
scenario of the ParameterizedAutoAddMasterTest test.  In addition,
a bit of extra delay is now added between writing batches of test rows.

Change-Id: Id0f0005b8f3384ad05eee6bf65e17a0e4ed289f1
Reviewed-on: http://gerrit.cloudera.org:8080/21935
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Attila Bukor <abukor@apache.org>