[dynamic_multi_master-test] add SKIP_IF_SLOW_NOT_ALLOWED
This patch adds SKIP_IF_SLOW_NOT_ALLOWED() macro into appropriate
scenarios in the dynamic_multi_master-test.  The criteria to tag a
scenario as 'slow' was the evidence of requiring more than 3 seconds
to complete on a more-or-less contemporary machine (DEBUG build).

The motivation for this changelist is to address the review feedback
for one of the prior patches [1].

[1] review URL: http://gerrit.cloudera.org:8080/21935
    gerrit ID : Id0f0005b8f3384ad05eee6bf65e17a0e4ed289f1
    git hash  : dbb6d5320ad60016355d0c63b6dae1650ba4e414

Change-Id: Ib255caf1ae33ebf9df9e71d9ea784d6e7d350ed9
Reviewed-on: http://gerrit.cloudera.org:8080/21943
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Marton Greber <greber.mx@gmail.com>