KUDU-613: Fix minor metrics related bug
When upserting in the protected segment, the state
of the handle isn't updated properly (the variable
in_protected_segment). This variable is used later
when releasing the handle to determine which segment's
metrics to update. This patch fixes this bug and adds
appropriate test coverage to verify the right metrics
are being updated.

Change-Id: I0d2fd44169e1abe642c11e16e4b87aaa971b2d88
Reviewed-on: http://gerrit.cloudera.org:8080/21956
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>