KUDU-3623 do not crash on invalid input for 'kudu table copy'
This patch adds code to handle invalid input for the --predicates flag
of the 'kudu table copy' tool.  Now, instead of crashing on any
non-expected input, the tool exists gracefully and reports an actionable
error message.  With that, now it's easier to understand what was wrong,
correct typos/mistakes, and re-run the tool again, eventually achieving
the desired goal.

A new test is added to cover the newly introduced functionality.

This is a follow-up to 0afeddf9e530762e0e47beb7428982763715c746.

Change-Id: I96ee4c38f3074b44ce89ac1b18784058e9148daf
Reviewed-on: http://gerrit.cloudera.org:8080/22009
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>