[cfile] 'null bitmap' --> 'non-null bitmap'
The corresponding bitmaps have always had the semantics that a set bit
indicated a non-null entry.  Thus naming any of those as a 'null bitmap'
is quite misleading.  Also, since changelist [1] renamed a subset of
entities already, now it's even more confusing.

I think it's never too late to add more clarity, so this patch renames
the rest of the entities in the code and also updates the documentation.

This patch doesn't contain any functional modifications.

This is a follow-up to [1].

[1] https://github.com/apache/kudu/commit/ffef96b26

Change-Id: Ib298b5b6bca786dc69928b960c535f95f177d604
Reviewed-on: http://gerrit.cloudera.org:8080/22075
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>