[build-and-test] run 'gradle clean' before building Java
When updating Java dependencies in changelists [1] and [2], I noticed
that pre-commit builds were failing fail because the copyDistTestJars
task wouldn't copy all the necessary JARs to into the staging directory
for dist-test since gradle would assume the target is up-to-date.
I think that was due to the fact that the same workspace was re-used
between different builds, and no 'clean' task was run.

This changelist is to address the issue.  To avoid incurring the penalty
of rebuilding Java bits from scratch, this patch enables gradle's build
cache. As per [3], this is exactly the use-case to benefit from gradle's
caching.  In addition, this patch re-enables parallel builds as well.

[1] https://github.com/apache/kudu/commit/b009144cd
[2] https://github.com/apache/kudu/commit/150ec7ff5
[3] https://docs.gradle.org/current/userguide/part6_gradle_caching.html

Change-Id: Ief3d2f07dfcf8746aa69f870c319599c92e098dc
Reviewed-on: http://gerrit.cloudera.org:8080/22144
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
(cherry picked from commit 91cb9da22c9437db2e8b4b0f2793aca081afddc1)
Reviewed-on: http://gerrit.cloudera.org:8080/22225