[build-support-test] remove TEST_TMPDIR with its content
In case of concurrent build/test tasks, it makes sense to create
per-task TEST_TMPDIR directory because of the logic in the
jenkins/build-and-test.sh script that introspects its content.
When doing so, it make sense to remove the directory along with its
content, not just the content under the directory.  Otherwise,
an empty TEST_TMPDIR directory would be left behind.  It's not an
issue when TEST_TMPDIR is the same for every build/task, but with
custom, per-build TEST_TMPDIR, there would be too many empty directories
left behind.

With this patch, the script is amended to address that: as a part
of clean-up, it now removes the TEST_TMPDIR directory along with its
content, not just the content of the directory.

Change-Id: I1d79f9482575fecba2e81c2d5fe85cbe25afd6b5
Reviewed-on: http://gerrit.cloudera.org:8080/22150
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 8027299167f6f4f0fe586b6c70829054aae31dab)
Reviewed-on: http://gerrit.cloudera.org:8080/22152