KUDU-613: Remove use of vectors under lock
Currently, the implementation of the SLRU cache returns
a vector of temporarily evicted entries to be reinserted
later. There are performance concerns with this approach
as it allocates and deallocates memory under the lock.

This patch refactors the code to not use a vector to deal
with the temporarily evicted entries.

Change-Id: I8f6a019c7c033c6c3dfef59b0f037e78f8264e68
Reviewed-on: http://gerrit.cloudera.org:8080/22181
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 33fd5e482b4854d23247af903c4263afc9931e59)
Reviewed-on: http://gerrit.cloudera.org:8080/22195