KUDU-613: Micro-optimization in Erase()
Prior to this patch, the Erase() function on the shard level
did not have a return value. The Erase() function on the
ShardPair level called both shard-level functions.

This patch changes the shard-level Erase() function to return
a bool to indicate whether an entry was erased. This allows
the ShardPair level function to only call Erase() on the
protected shard when the probationary shard does not find and
erase the appropriate entry.

Change-Id: I9941b671f00d5f4dd71b59a77b45f9cdbd79b367
Reviewed-on: http://gerrit.cloudera.org:8080/22229
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 5151270ea9d88bc40ad5ca2344c97f1eed5fadc9)
Reviewed-on: http://gerrit.cloudera.org:8080/22234