KUDU-3591 a few extra options to speed up RocksDB bootstrapping
Even with changelist [1], RocksDB-specific flavors of the
FsManagerTestBase.TestAddRemoveDataDirsFuzz scenario still fail from
time to time even in non-sanitized DEBUG builds due to running out of
time allowance.

This patch introduces a couple of extra flags for the "logr" block
manager to use in test scenarios like that and tweaks already existing
flags to speed up bootstrapping of the embedded RocksDB instance.
Also, the number of iterations for the scenario has decreased 4 times.

As a quick validation of the scenario's runtime before and after this
patch, I ran the following multiple times:

  time KUDU_ALLOW_SLOW_TESTS=1 ./bin/fs_manager-test --gtest_filter='BlockManagerTypes/FsManagerTestBase.TestAddRemoveDataDirsFuzz/4' --gtest_repeat=10 > /dev/null 2>&1

There is high variation in run times due to the stochastic nature
of the test scenario, but after running it many times I found that the
numbers below provide a good sense of the corresponding run times.

Before:
  real    1m1.831s
  user    0m36.494s
  sys     0m16.228s

After:
  real    0m10.237s
  user    0m5.945s
  sys     0m2.872s

This is a follow-up to [1].

[1] https://github.com/apache/kudu/commit/4116941c0

Change-Id: If95282d0e5ef3f5c1bd20e04dd24ac0a14374cee
Reviewed-on: http://gerrit.cloudera.org:8080/22232
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>