[tools] Run range aware cluster rebalance with multiple tables
Previously when rebalancing cluster with kudu command line tool, the
`--enable_range_rebalancing` flag requires the `--tables` flag with exactly one
table specified.

This commit removes this restriction, allowing multiple tables to be passed to
the `--tables` flag when range rebalancing is enabled. Among them, range
partitioned tables will be rebalanced with ranges considered, while others are
rebalanced as usual. If the `--tables` flag is not set, all tables in the
cluster will be rebalanced.

This commit also optimizes the replica distribution details when range
rabalancing is applied to multiple tables with mixed partition schemas.

Change-Id: I7f6995496e6c28e6a3a8fa4b6f0e8efd05335267
Reviewed-on: http://gerrit.cloudera.org:8080/22239
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>