KUDU-613: Make segment metrics more relevant
This patch removes the segment-level lookup metrics
along with the upgrade and downgrade counter metrics.

It instead adds histogram metrics to keep track of
the stats of upgrades and downgrades per entry. The
total number of upgrades and downgrades for the cache
can be derived from these new metrics using the field
'total_count'.

Counters to keep track of upgrades and downgrades
are also added on the handle level.

Change-Id: Id44828be41a19593e6808debf7e9ba8e1fc4dcca
Reviewed-on: http://gerrit.cloudera.org:8080/22253
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>