[thirdparty] fix building LLVM/CLANG 11.0.0 with GCC13
GCC13 (G++13) is the default system compiler on contemporary Linux OS
flavors such as Ubuntu 24.04 LTS.  With this patch it's now possible
to compile LLVM/CLANG in Kudu's thirdparty on Ubuntu 24.04 and 24.04.1
LTS.

This patch picks up a few relevant updates from the LLVM project's
upstream repository [1] with minor tweaks to adapt to the change in
the source files layout that happened in LLVM/CLANG 12.  The updates
prevent the leakage of the parts of the standard library installed
on the build machine into the components of the LLVM/CLANG
that are supposed to be completely freestanding and standalone.
I guess it should help not only with making it possible to build
LLVM/CLANG 11 with GCC13, but also address the root cause of the issues
we have hit on SLES15 and elsewhere in the codegen that's reported in
a few Jira items such as KUDU-3479 and KUDU-3545 [2][3].
However, there might be some extra patches beyond the ones already
included.  Anyway, I haven't verified this hypothesis yet.

We definitely need to upgrade from LLVM 11.0.0 to LLVM 19.1.6 and I
suspect it should help us to address [2] and [3] along with other
updates and fixes.  However, it's a separate task on its own,
while this patch focuses on making the current version of LLVM/CLANG
in the Kudu's thirdparty buildable on recent Linux OS releases
(such as Ubuntu 24.04 LTS).

I also snuck in an update to skip building compiler-rt tests since
they aren't run in the scope of building LLVM/CLANG as a thirdparty
component of the Apache Kudu project.

[1] https://github.com/llvm/llvm-project
[2] https://issues.apache.org/jira/browse/KUDU-3479
[3] https://issues.apache.org/jira/browse/KUDU-3545

Change-Id: I9b8282c85bbac5f25eae885e4b3ec183104bc540
Reviewed-on: http://gerrit.cloudera.org:8080/22279
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Ashwani Raina <araina@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>