[build-and-test] TEST_TMPDIR default setting for Jenkins
In case of Jenkins builds, TEST_TMPDIR must be created per-build and
cleaned up upon exiting.  This patch updates the default setting
for the TEST_TMPDIR in the context of a Jenkins job, and properly
cleans up the TEST_TMPDIR directory upon exit.

This is a follow-up to 8027299167f6f4f0fe586b6c70829054aae31dab.

Change-Id: Ie6115d68526d90f8ef627b7d340f2f11a4e3d3cc
Reviewed-on: http://gerrit.cloudera.org:8080/22308
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>