[docs] clarify on multi-row transaction support in known issues
It's quite surprising to see the 'Known Issues' document still stating
that multi-row transactions are not at all supported in Kudu while the
support for multi-row transactions across multiple tables has been
present in Kudu for INSERT and INSERT_IGNORE operations for a few years.

This patch updates the known_issues.adoc on this particular topic.

Change-Id: Iac6652549f2593e041200ed1cdf7b2771c052b70
Reviewed-on: http://gerrit.cloudera.org:8080/22321
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>