[compaction] Release delta blocks memory early to avoid more consumption.
MajorDelta compaction relies on object scope to release the memory
used to store delta blocks that are read during the PrepareBatch phase
of the compaction operation. The use of delta blocks is till the point
when REDO mutations are applied to base data and corresponding UNDO
mutations are created for each REDO mutation.

This patch releases the memory as soon as compaction is done using
delta blocks. This will avoid unnecessary pressure on memory for the
rest of the compaction operation. This is also a precursor to MajorDelta
chunked compaction work.

Change-Id: I0c1eadcb8ad735fc92ac77f554934e606e4119ae
Reviewed-on: http://gerrit.cloudera.org:8080/22338
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>