[util] omit '-b' option from lsof's invocations
I haven't checked for the reason behind this, but at least with kernel
6.8.0 on Ubuntu 24.04.1 LTS, lsof doesn't list TCP sockets when running
with '-b' option (lsof version 4.95.0) like below:
  lsof -wbnP -Fpfn -a -i 4TCP -p <pid>

Both TCP and UDP sockets are listed as expected when not adding
the '-b' option into the lsof's command line.

Removing the '-b' option shouldn't be a big deal since nothing
in the upstream CI pipeline runs on machines with NFS mounts or other
types of stale mounts that would block lstat(), readlink(), or stat()
system calls [1].  In any case, even if a process running as a part
of Kudu test scaffolding opens filesystem objects on a NFS mount,
the result of the test isn't quite predictable when the NFS mount
stalls, regardless of the lsof's behavior.

[1] https://www.man7.org/linux/man-pages/man8/lsof.8.html

Change-Id: I73490a8f73728434292c5ee6f175eed63523ffd4
Reviewed-on: http://gerrit.cloudera.org:8080/22380
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>