[util] improve MaintenanceManagerTest.TestRunningInstances
This patch introduces a ScopedCleanup instance to unregister maintenance
operations registered by MaintenanceManagerTest.RunningInstances if
ASSERT_EVENTUALLY triggers.  Also, the scenario is now more resilient to
scheduler anomalies.  The motivation for this patch was a report
on test failures [1] with the following TSAN warning due to CHECK()
in MaintenanceOp::~MaintenanceOp():

  WARNING: ThreadSanitizer: data race on vptr (ctor/dtor vs virtual call)

[1] http://dist-test.cloudera.org/job?job_id=jenkins-slave.1738131961.3963207

Change-Id: I95f34e7cb8d37a0da4445012de40abfb2fe6f00e
Reviewed-on: http://gerrit.cloudera.org:8080/22427
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Kudu Jenkins