[site] Remove thirdparty resource references
This commit updates Kudu website by removing any resource reference
to any thirparty hosted website and replacing them with files in the
local repo. This is needed to ensure the site works as expected once
the CSP blocking is in effect by Apache from March 1.

This commit also reverts commit 55b58c as the approach taken did not
work as expected.

Also updated site_tool script to use the correct command for html-proofer
and removed the verbose option as it is not present anymore.

Change-Id: I5971a406968bc69fc21bbd05618d85e702210539
Reviewed-on: http://gerrit.cloudera.org:8080/22472
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>