[common] small clean-up on PartialRow
It's possible to use 'if constexpr' to enable compile-time optimisations
in template code since switching to C++17.  This patch updates the
implementations of KuduPartialRow::{Set,SetSliceCopy}() methods to use
compile-time evaluation of 'type' and 'physical_type'.  I also added
static_assert() and PREDICT_FALSE() in the modified code where it's
appropriate, and did a few other style-related updates.

Change-Id: I1b164eb7ec7f010e9228175a026aa63726a909c6
Reviewed-on: http://gerrit.cloudera.org:8080/22480
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>