KUDU-3641 fix flaky TestNewLeaderCantResolvePeers (take 2)
I noticed that even after [1] the TestNewLeaderCantResolvePeers was
still failing in about once in 60 runs [2], so I took a closer look.

It turns out StartElection() doesn't trigger a re-election if it
arriving at the current Raft leader.  To address that, this patch
replaces StartElection() back with LeaderStepDown() but with new
target leader being the tablet replica at the third tablet server.

This is a follow-up to [1].

[1] https://github.com/apache/kudu/commit/6c77ec875
[2] http://dist-test.cloudera.org:8080/test_drilldown?test_name=raft_consensus_election-itest

Change-Id: I3bee924353079f7c8bfab6d0d5a6367bd1ee243e
Reviewed-on: http://gerrit.cloudera.org:8080/22516
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>
(cherry picked from commit 8ce91854a3ae749ea02c45096dfff4b877050a82)
Reviewed-on: http://gerrit.cloudera.org:8080/22517