[consensus] fix result status handling
This patch addresses sloppy result handling of functions/methods
that return Status under the src/kudu/consensus directory.

In addition, this patch introduces a new unsafe runtime flag
--raft_allow_committed_pending_index_gap that's now used by the
TsRecoveryITestDeathTest.RecoverFromOpIdOverflow test scenario only.
The test scenario was reliant on the sloppy result status handling
of the call to PendingRounds::SetInitialCommittedOpId() in
RaftConsensus::Start() that's been addressed in this patch.  It's
the path of least resistance compared with updating test scaffolding
and the test scenario itself, but it also has a benefit of providing
a flag to enable the former result status handling approach in
RaftConsensus::Start() when setting last committed OpId for a tablet's
Raft consensus implementation.  Frankly, I don't think anybody will ever
need to re-enable the original bogus behavior of ignoring non-OK result
status, but it doesn't hurt to have such an option just in case
if I'm missing something here.

Change-Id: I6db113c58c6b453bf3cd290ee0dfdd184b63b999
Reviewed-on: http://gerrit.cloudera.org:8080/22646
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>